fix upload img at review create form

+++ frontend/models/ImageUpload.php
@@ -4,34 +4,34 @@ namespace frontend\models;
 
 use Yii;
 use yii\base\Model;
-
+use yii\web\UploadedFile;
 
 class ImageUpload extends Model{
-    public $image;
+    public $img;
 
     public function rules()
     {
         return [
-            [['image'] , 'required'],
-            [['image'], 'file', 'extensions' => 'jpg, png']
+            [['img'] , 'required'],
+            [['img'], 'file', 'extensions' => 'jpg, png']
         ];
     }
 
     public function uploadFile($file, $currentImage)
     {
-        if ($this->validate())
-        {
-            $this->image = $file;
-            if(file_exists(Yii::getAlias('@web') . 'uploads/' . $currentImage))
+            $this->img = $file;
+
+
+            if(!empty($currentImage) && $currentImage != null && file_exists(Yii::getAlias('@web') . 'uploads/' . $currentImage))
             {
                 unlink(Yii::getAlias('@web') . 'uploads/' . $currentImage);
             }
 
             $filename = strtolower(md5(uniqid($file->baseName)) . '.' . $file->extension);
+            $this->img->saveAs( Yii::getAlias('@web') . 'uploads/' . $filename);
 
-            $file->saveAs(Yii::getAlias('@web' . 'uploads/' . $filename));
-            return $file->name;
-        }
+            return $filename;
+
 
     }
 }
